Clear what Xcode 27's clang adds, and name CC#64 where it is emitted
Xcode 27 ships clang 21, which reports warnings the compiler behind
e34d19a did not. Nothing here changes behaviour.

-Wdouble-promotion in six places, every one a float widening to double
that was already intended and is now written down. output_filter's
`process` accounted for most of the noise by itself: it arrives through
tone_generator.hpp, so the two lines seeding `in_left`/`in_right` fired
in twelve translation units. The rest are equalizer's per-sample pair,
three `ramp.step()` readings in control_ramp_tests -- beside a fourth
that already carried the cast -- the float target handed to Catch2's
`WithinAbs(double, double)`, and note_oracle's metric accumulation.

-Wunused-const-variable on `cc_sustain`, which wanted the decompile
rather than a deletion. CC#64 is handled everywhere it should be:
`cc64_hold_damper @ 180065e50` splits on tone header byte 0x0d, keeping
the raw pedal value on half-damper tones and quantising to 0/0x7f
elsewhere, and pitch_chain reproduces its release scaling term for term.
The constant was dead in modulators.cpp alone, orphaned by 11e6232 when
the bank-wide CC#64 modulator moved onto the per-instrument global zones.

So it returns where it is emitted, as `cc_damper` -- pairing with
`cc_soft` as the other pedal, and matching what the engine calls this
controller in every other file. bank_builder was writing all five of its
source words as bare literals; the curve shapes and controller numbers
now live in modulator_names.hpp and both halves of the export read the
same names. A zone modulator replaces a bank default only when the two
source words match exactly, so the two files agreeing is load-bearing
rather than tidiness.

The export was verified byte-identical across this change when it was
first written -- the whole bank, 0364f11d over 69.3 MB, with the same 104
half-damper and 2201 filter-envelope instruments either way. The casts
are value-preserving widenings and the rest is renaming, so there is
nothing here a render can see.

Suite with the current fixtures: 235 passed, 4 failed, 1 skipped over
2,635,368 assertions, and the player preset builds clean.

+++ include/tabulasonora/output_filter.hpp
@@ -63,8 +63,8 @@ public:
     /// path would have to carry that difference before this could serve it.
     [[nodiscard]] std::pair<float, float> process(float left, float right) noexcept
     {
-        const double in_left = left;
-        const double in_right = right;
+        const double in_left = static_cast<double>(left);
+        const double in_right = static_cast<double>(right);
 
         // The allpass, per channel: the state takes the input less the fed-back state, and the
         // output leads with the coefficient. Half a sample of delay, which is the midpoint the
